Add missing gender field in new customer form
Customers -> New customer was missing the gender field. Filling this
form out resulted in a QueryException:
Integrity constraint violation:
19 NOT NULL constraint failed: shop_customers.gender
Fixing this by adding the gender field to this form. This is similar to
PR #64 which fixed a similar issue when adding a customer while creating
a new order (Orders -> New order -> Adding a customer there)